fixed scanning and compaction

+++ stream_compaction/common.cu
@@ -24,7 +24,7 @@ namespace StreamCompaction {
          */
         __global__ void kernMapToBoolean(int n, int *bools, const int *idata) {
             // TODO
-            int index = threadIdx.x + (blockIdx.x * blockDim.x);
+            size_t index = threadIdx.x + (blockIdx.x * blockDim.x);
             if (index >= n) {
                 return;
             }
@@ -38,7 +38,7 @@ namespace StreamCompaction {
         __global__ void kernScatter(int n, int *odata,
                 const int *idata, const int *bools, const int *indices) {
             // TODO
-            int index = threadIdx.x + (blockIdx.x * blockDim.x);
+            size_t index = threadIdx.x + (blockIdx.x * blockDim.x);
             if (index >= n) {
                 return;
             }
